+++ src/number.core.ts
@@ -401,11 +401,11 @@ export function validateNumber(values: readonly unknown[], {
 
 }: NumberShape, {
 
-	placeholder=false
+	model=false
 
 }: {
 
-	placeholder?: boolean
+	model?: boolean
 
 }={}): undefined | Trace {
 
@@ -447,7 +447,7 @@ export function validateNumber(values: readonly unknown[], {
 			|| `expected values in [${allowed.join(", ")}]`
 		),
 
-		"{hasValue}": placeholder || group(matching, group =>
+		"{hasValue}": model || group(matching, group =>
 			hasValue === undefined || hasValue.every(v => group.includes(v))
 			|| `expected values to include [${hasValue.join(", ")}]`
 		)